Refactor: move stores to domain directories
